feat: stip
