+++ tests/e2e_tests/test_end_to_end.py
@@ -405,21 +405,6 @@ def test_add_upload_workflow_with_alert_trigger(browser: Page):
         raise
 
 
-def test_start_with_keep_db(browser: Page, setup_page_logging, failure_artifacts):
-    # Navigate to signin page
-    browser.goto("http://localhost:3001/signin")
-    # Fill in credentials
-    browser.get_by_placeholder("Enter your username").fill("keep")
-    browser.get_by_placeholder("Enter your password").fill("keep")
-    # Click sign in and wait for navigation
-    browser.get_by_role("button", name="Sign in").click()
-    try:
-        browser.wait_for_url("http://localhost:3001/incidents", timeout=10000)
-    except Exception:
-        save_failure_artifacts(browser)
-        raise
-
-
 def test_provider_deletion(browser: Page):
     log_entries = []
     setup_console_listener(browser, log_entries)
